CREATED:sendEmail function
