+++ app.js
@@ -1,33 +1,33 @@
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
-const nodemailer = require("nodemailer");
+// const nodemailer = require("nodemailer");
 require("dotenv").config();
-const { UKR_NET_EMAIL, UKR_NET_PASSWORD } = process.env;
+// const { UKR_NET_EMAIL, UKR_NET_PASSWORD } = process.env;
 
-const nodemailerConfig = {
-  host: "smtp.ukr.net",
-  port: 465,
-  secure: true,
-  auth: {
-    user: UKR_NET_EMAIL,
-    pass: UKR_NET_PASSWORD,
-  },
-};
+// const nodemailerConfig = {
+//   host: "smtp.ukr.net",
+//   port: 465,
+//   secure: true,
+//   auth: {
+//     user: UKR_NET_EMAIL,
+//     pass: UKR_NET_PASSWORD,
+//   },
+// };
 
-const transport = nodemailer.createTransport(nodemailerConfig);
+// const transport = nodemailer.createTransport(nodemailerConfig);
 
-const email = {
-  from: UKR_NET_EMAIL,
-  to: "popovaira969@gmail.com",
-  subject: "Please,verify email",
-  html: "<p>Please,verify email</p>",
-};
+// const email = {
+//   from: UKR_NET_EMAIL,
+//   to: "popovaira969@gmail.com",
+//   subject: "Please,verify email",
+//   html: "<p>Please,verify email</p>",
+// };
 
-transport
-  .sendMail(email)
-  .then(() => console.log("Email send success"))
-  .catch((error) => console.log(error.message));
+// transport
+//   .sendMail(email)
+//   .then(() => console.log("Email send success"))
+//   .catch((error) => console.log(error.message));
 
 const contactsRouter = require("./routes/api/contacts");
 const authRouter = require("./routes/api/auth");
